Initial workings communication with mastery grid
The parser needs major rewriting for legibility and all parts need logging.

To actually populate (and later refresh) the database a number of calls to the API equal to the number of students are needed (at the time of writing > 150 for TDT4100. This means some kind of periodic update mechanism is needed instead on a per query basis. Maybe with the ability to force a query. Possible pitfall might actually be the server thinking we are DOS-ing.